chore: code organization for trace types
